chore: rename env/generate action name
- env/generate -> file/updateEnv
