+++ packages/fx-core/resource/package.nls.json
@@ -1131,13 +1131,13 @@
   "driver.tools.summary.func.installed": "Azure Functions Core Tools is installed.",
   "driver.tools.summary.dotnet.installedWithPath": ".NET Core SDK is installed at %s.",
   "driver.tools.summary.dotnet.installed": ".NET Core SDK is installed.",
-  "driver.env.error.invalidParameter": "Following parameter is missing or invalid for %s action: %s.",
-  "driver.env.error.unhandledError": "Unhandled error happened in %s action: %s",
-  "driver.env.description": "generates environment variables to .env file",
-  "driver.env.summary.default": "The environment variables has been generated successfully to the .env file of '%s' environment.",
-  "driver.env.summary.withTarget": "The environment variables has been generated successfully to %s.",
-  "driver.env.progressBar.title": "env/generate",
-  "driver.env.progressBar.generate": "Generating the environment variables to .env file.",
+  "driver.file.error.invalidParameter": "Following parameter is missing or invalid for %s action: %s.",
+  "driver.file.error.unhandledError": "Unhandled error happened in %s action: %s",
+  "driver.file.description": "generates environment variables to .env file",
+  "driver.file.summary.default": "The environment variables has been generated successfully to the .env file of '%s' environment.",
+  "driver.file.summary.withTarget": "The environment variables has been generated successfully to %s.",
+  "driver.file.progressBar.title": "file/updateEnv",
+  "driver.file.progressBar.generate": "Generating the environment variables to .env file.",
   "driver.env.error.folderNotExist": "Folder %s does not exist.",
   "driver.deploy.error.unknownError": "Unknown error happened: %s",
   "driver.deploy.notice.deployAcceleration": "Deploying to Azure App Service takes a long time. Consider referring to this document to optimize your deployment:",
