+++ packages/fx-core/src/component/driver/file/appsettingsGenerate.ts
