+++ packages/fx-core/src/component/driver/file/error/invalidParameterUserError.ts
@@ -6,7 +6,7 @@ import { UserError } from "@microsoft/teamsfx-api";
 import { getDefaultString, getLocalizedString } from "../../../../common/localizeUtils";
 
 const errorCode = "InvalidParameter";
-const messageKey = "driver.env.error.invalidParameter";
+const messageKey = "driver.file.error.invalidParameter";
 
 export class InvalidParameterUserError extends UserError {
   constructor(actionName: string, invalidParameters: string | string[], helpLink: string) {
