+++ packages/fx-core/src/component/driver/file/interface/generateAppsettingsArgs.ts
