+++ packages/fx-core/src/component/driver/file/interface/generateEnvArgs.ts
