+++ packages/fx-core/src/component/driver/file/updateEnv.ts
@@ -20,12 +20,12 @@ import { InvalidParameterUserError } from "./error/invalidParameterUserError";
 import { UnhandledSystemError } from "./error/unhandledError";
 import { GenerateEnvArgs } from "./interface/generateEnvArgs";
 
-const actionName = "env/generate";
-const helpLink = "https://aka.ms/teamsfx-actions/env-generate";
+const actionName = "file/updateEnv";
+const helpLink = "https://aka.ms/teamsfx-actions/file-updateEnv";
 
 @Service(actionName) // DO NOT MODIFY the service name
-export class GenerateEnvDriver implements StepDriver {
-  description = getLocalizedString("driver.env.description");
+export class UpdateEnvDriver implements StepDriver {
+  description = getLocalizedString("file/updateEnv.description");
 
   @hooks([addStartAndEndTelemetry(actionName, actionName)])
   public async run(
@@ -59,7 +59,7 @@ export class GenerateEnvDriver implements StepDriver {
     summaries: string[];
   }> {
     const progressHandler = context.ui?.createProgressBar(
-      getLocalizedString("driver.env.progressBar.title"),
+      getLocalizedString("file/updateEnv.progressBar.title"),
       1
     );
 
@@ -68,7 +68,7 @@ export class GenerateEnvDriver implements StepDriver {
 
       this.validateArgs(args);
 
-      await progressHandler?.next(getLocalizedString("driver.env.progressBar.generate"));
+      await progressHandler?.next(getLocalizedString("file/updateEnv.progressBar.generate"));
 
       if (args.target) {
         const target = this.getAbsolutePath(args.target, context.projectPath);
@@ -83,7 +83,9 @@ export class GenerateEnvDriver implements StepDriver {
 
         return {
           output: new Map<string, string>(),
-          summaries: [getLocalizedString("driver.env.summary.withTarget", path.normalize(target))],
+          summaries: [
+            getLocalizedString("file/updateEnv.summary.withTarget", path.normalize(target)),
+          ],
         };
       } else {
         const state = this.loadCurrentState();
@@ -92,7 +94,7 @@ export class GenerateEnvDriver implements StepDriver {
 
         return {
           output: new Map(Object.entries(args.envs)),
-          summaries: [getLocalizedString("driver.env.summary.default", state.TEAMSFX_ENV)],
+          summaries: [getLocalizedString("file/updateEnv.summary.default", state.TEAMSFX_ENV)],
         };
       }
     } catch (error) {
