+++ packages/fx-core/src/core/FxCore.ts
@@ -142,8 +142,8 @@ import "../component/driver/script/dotnetBuildDriver";
 import "../component/driver/script/npmBuildDriver";
 import "../component/driver/script/npxBuildDriver";
 import "../component/driver/tools/installDriver";
-import "../component/driver/env/generate";
-import "../component/driver/env/appsettingsGenerate";
+import "../component/driver/file/updateEnv";
+import "../component/driver/file/appsettingsGenerate";
 import "../component/driver/botFramework/createOrUpdateBot";
 import { settingsUtil } from "../component/utils/settingsUtil";
 import { DotenvParseOutput } from "dotenv";
