+++ templates/scenarios/js/non-sso-tab-default-bot/teamsfx/app.local.yml
@@ -23,7 +23,7 @@ provision:
       description: ""
 
 configureApp:
-  - uses: env/generate # Generate env to .env file
+  - uses: file/updateEnv # Generate env to .env file
     with:
       envs:
         TAB_DOMAIN: localhost:53000
